// web-client/src/api/webApiClient.ts
export class RecipeCreateDto implements IRecipeCreateDto {
    ingredients?: RecipeIngredientCreateDto[] | null;
    description?: string | null;
    name?: string | null;
    categoryId?: string;

    constructor(data?: IRecipeCreateDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            if (Array.isArray(_data["ingredients"])) {
                this.ingredients = [] as any;
                for (let item of _data["ingredients"])
                    this.ingredients!.push(RecipeIngredientCreateDto.fromJS(item));
            }
            else {
                this.ingredients = null as any;
            }
            this.description = _data["description"] !== undefined ? _data["description"] : null as any;
            this.name = _data["name"] !== undefined ? _data["name"] : null as any;
            this.categoryId = _data["categoryId"] !== undefined ? _data["categoryId"] : null as any;
        }
    }

    static fromJS(data: any): RecipeCreateDto {
        data = typeof data === 'object' ? data : {};
        let result = new RecipeCreateDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        if (Array.isArray(this.ingredients)) {
            data["ingredients"] = [];
            for (let item of this.ingredients)
                data["ingredients"].push(item ? item.toJSON() : null as any);
        }
        data["description"] = this.description !== undefined ? this.description : null as any;
        data["name"] = this.name !== undefined ? this.name : null as any;
        data["categoryId"] = this.categoryId !== undefined ? this.categoryId : null as any;
        return data;
    }
}

export interface IRecipeCreateDto {
    ingredients?: RecipeIngredientCreateDto[] | null;
    description?: string | null;
    name?: string | null;
    categoryId?: string;
}
// ...
export class RecipeIngredientCreateDto implements IRecipeIngredientCreateDto {
    ingredientId?: string;
    quantity?: string | null;

    constructor(data?: IRecipeIngredientCreateDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.ingredientId = _data["ingredientId"] !== undefined ? _data["ingredientId"] : null as any;
            this.quantity = _data["quantity"] !== undefined ? _data["quantity"] : null as any;
        }
    }

    static fromJS(data: any): RecipeIngredientCreateDto {
        data = typeof data === 'object' ? data : {};
        let result = new RecipeIngredientCreateDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["ingredientId"] = this.ingredientId !== undefined ? this.ingredientId : null as any;
        data["quantity"] = this.quantity !== undefined ? this.quantity : null as any;
        return data;
    }
}

export interface IRecipeIngredientCreateDto {
    ingredientId?: string;
    quantity?: string | null;
}
// ...
export class RecipeShortDto implements IRecipeShortDto {
    id?: string;
    name?: string | null;
    description?: string | null;
    categoryId?: string;
    createdAt?: Date;

    constructor(data?: IRecipeShortDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"] !== undefined ? _data["id"] : null as any;
            this.name = _data["name"] !== undefined ? _data["name"] : null as any;
            this.description = _data["description"] !== undefined ? _data["description"] : null as any;
            this.categoryId = _data["categoryId"] !== undefined ? _data["categoryId"] : null as any;
            this.createdAt = _data["createdAt"] ? new Date(_data["createdAt"].toString()) : null as any;
        }
    }

    static fromJS(data: any): RecipeShortDto {
        data = typeof data === 'object' ? data : {};
        let result = new RecipeShortDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id !== undefined ? this.id : null as any;
        data["name"] = this.name !== undefined ? this.name : null as any;
        data["description"] = this.description !== undefined ? this.description : null as any;
        data["categoryId"] = this.categoryId !== undefined ? this.categoryId : null as any;
        data["createdAt"] = this.createdAt ? this.createdAt.toISOString() : null as any;
        return data;
    }
}

export interface IRecipeShortDto {
    id?: string;
    name?: string | null;
    description?: string | null;
    categoryId?: string;
    createdAt?: Date;
}

export class RecipeShortDtoPageResult implements IRecipeShortDtoPageResult {
    items?: RecipeShortDto[] | null;
    page?: number;
    pageSize?: number;
    totalCount?: number;

    constructor(data?: IRecipeShortDtoPageResult) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            if (Array.isArray(_data["items"])) {
                this.items = [] as any;
                for (let item of _data["items"])
                    this.items!.push(RecipeShortDto.fromJS(item));
            }
            else {
                this.items = null as any;
            }
            this.page = _data["page"] !== undefined ? _data["page"] : null as any;
            this.pageSize = _data["pageSize"] !== undefined ? _data["pageSize"] : null as any;
            this.totalCount = _data["totalCount"] !== undefined ? _data["totalCount"] : null as any;
        }
    }

    static fromJS(data: any): RecipeShortDtoPageResult {
        data = typeof data === 'object' ? data : {};
        let result = new RecipeShortDtoPageResult();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        if (Array.isArray(this.items)) {
            data["items"] = [];
            for (let item of this.items)
                data["items"].push(item ? item.toJSON() : null as any);
        }
        data["page"] = this.page !== undefined ? this.page : null as any;
        data["pageSize"] = this.pageSize !== undefined ? this.pageSize : null as any;
        data["totalCount"] = this.totalCount !== undefined ? this.totalCount : null as any;
        return data;
    }
}

export interface IRecipeShortDtoPageResult {
    items?: RecipeShortDto[] | null;
    page?: number;
    pageSize?: number;
    totalCount?: number;
}